+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from "react-router-dom";
 import router from "./router/router";
-import GlobalFallback from "./components/layout/fallback/GlobalFallback";
+import GlobalFallback from "./pages/fallback/GlobalFallback";
 
 function App() {
   return <RouterProvider router={router} fallbackElement={<GlobalFallback />} />;
